requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20
pydantic==2.12.5
ortools==9.11.4210
bcrypt==4.2.1
pyjwt==2.10.1